refactor:이전 가비지 함수들 모두 정리 완료.
