+++ src/main/java/kr/codesquad/CustomException.java
@@ -1,7 +1,7 @@
 package kr.codesquad;
 
-public class CustomException extends Exception{
-    public CustomException (String moneyNotDividedClearly){
+public class CustomException extends Exception {
+    public CustomException(String moneyNotDividedClearly) {
         super(moneyNotDividedClearly);
     }
 }
