+++ src/main/java/kr/codesquad/LottoManager/LottoGenerator.java
@@ -1,10 +1,10 @@
 package kr.codesquad.LottoManager;
 
 import kr.codesquad.CustomException;
-import kr.codesquad.InputHandler;
+import kr.codesquad.Utility.InputHandler;
 import kr.codesquad.User.UserInput;
 import kr.codesquad.User.UserLotto;
-import kr.codesquad.Utility;
+import kr.codesquad.Utility.Utility;
 
 public class LottoGenerator extends UserInput{
     // 유틸리티나 인풋 핸들러들을 일일이 계속 선언해주고있는데, 개선 하고 싶음. user inputt
